# Remove call to Valid() during encoding to boost speed by about 13-18%