Closes #851

## Description

- [x] eliminate call to Valid() in the encoder
- [x] increase max limits in decoder to eliminate one reason for encoder to call Valid().

Thanks @turbolent for suggesting this optimization last week.

### Benchmark Comparisons

Speed improved by 13% to almost 18%.  Memory use is unchanged. 

```
name                                               old time/op    new time/op    delta
EncodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      148µs ± 1%     121µs ± 1%  -17.97%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4       4.33µs ± 1%    3.65µs ± 1%  -15.68%  (p=0.000 n=10+10)
EncodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4       3.22µs ± 1%    2.74µs ± 1%  -14.89%  (p=0.000 n=10+9)
EncodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4    7.04ms ± 1%    6.12ms ± 0%  -13.09%  (p=0.000 n=10+10)

------

name                                            old time/op    new time/op    delta
EncodeCBOR/link_v3_2392f05c3b72f235_167bytes-4    3.30µs ± 2%    2.72µs ± 1%  -17.52%  (p=0.000 n=10+9)
EncodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4    3.31µs ± 1%    2.75µs ± 2%  -17.04%  (p=0.000 n=9+10)
```
Benchmarks used Go 1.15.11 on linux_amd64.

### Caveats

:warning: WARNING:  Max limits for array size, etc. are increased -- if adversaries can cause excessively large CBOR data to be decoded, they can launch resource exhaustion attacks.

Max limit for nested level is set to MaxInt16 because MaxInt32 would allow nested levels that would crash Go because of the 1 GB stack size limit of Go.
______

For contributor use:

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels